requirements.txt
numpy
opencv-python
Pillow
apng
fastapi
uvicorn[standard]
python-multipart
